Lots of enhancements.
Configured a a default layout for the home page.  The controller now resolves,
with no other configuration changes to a tiles definition that includes
a header and a footer page.  The HTML is well formed.  Also the title is
parameterized in the tiles definition.

All of the Dojo still works as it did before.

Added some styling as well.  This properly comes from a style sheet. that is included
in the default tiles layout.